refactor(#54): bookReviewEntity.ts -> bookReview.entity.ts 리네이밍
